// src/app/models/ticket.model.ts
export class Ticket {
    TicketId: number = 0;
    PassengerName: string = '';
    PassengerEmail: string = '';
    PassengerPhone: string = '';
    Gender: string = '';
    Address: string = '';
    JourneyDate: string = '';
    CustomerId: number = 0;
    FlightId: number = 0;
    TotalFare: number = 0;
    Age: number = 18;
    Status: boolean = false;
}